+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -396,6 +396,12 @@ New checks
   Detects issues in accordance with MISRA C rule 3.2:
   "`Line-splicing shall not be used in // comments`".
 
+- New :doc:`misrac-rule-7p1
+  <clang-tidy/checks/misrac/rule-7p1>` check.
+
+  Detects issues in accordance with MISRA C rule 7.1:
+  "`Octal constants shall not be used`".
+
 - New :doc:`misrac-rule-8p2
   <clang-tidy/checks/misrac/rule-8p2>` check.